+++ ci4/app/Interfaces/ServicePort.php
@@ -4,6 +4,8 @@
  * @property string $name
  * @property int $port
  * @property int $targetPort
+ * @property string $healthCheckType
+ * @property string $healthCheckPath
  */
 interface ServicePort {